pyproject.toml
[tool.poetry]
name = "twitter-clone-api"
version = "0.1.0"
description = "Api Twitter Clone"
authors = ["RBxRobson <robsonnbarbosa0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
factory-boy = "^3.3.1"
gunicorn = "^23.0.0"
whitenoise = "^6.8.1"
django-rest-framework = "^0.1.0"
pillow = "^11.0.0"
djangorestframework-simplejwt = "^5.3.1"
drf-nested-routers = "^0.94.1"
pymysql = "^1.1.1"
markdown = "^3.7"
django-environ = "^0.11.2"
cryptography = "^44.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
